Fixing UI rendering under dark layer caused by light system implementation, excluding build from dir
