+++ .gitignore
@@ -1,6 +1,6 @@
 
 # Created by https://www.gitignore.io/api/osx,macos,linux,node,windows
-
+build
 ### OSX ###
 *.DS_Store
 .AppleDouble
